refactor: move wait section utils to wait section file
