+++ src/page-modules/assistant/details/trip-section/index.tsx
@@ -15,8 +15,8 @@ import { ColorIcon } from '@atb/components/icon';
 import { MessageBox } from '@atb/modules/situations';
 import { PageText, useTranslation } from '@atb/translations';
 import { InterchangeDetails, InterchangeSection } from './interchange-section';
-import { formatLineName, getPlaceName, type LegWaitDetails } from '../utils';
-import WaitSection from './wait-section';
+import { formatLineName, getPlaceName } from '../utils';
+import WaitSection, { type LegWaitDetails } from './wait-section';
 
 export type TripSectionProps = {
   isFirst: boolean;
